:bug: Open local folders through the file input path.
The folder control now uses label-backed directory input activation instead of the programmatic picker path. Folder normalization and preview graph building live in the project-folder service so the page only coordinates import state.
