+++ web/src/services/graph-import-service.ts
@@ -1,22 +1,5 @@
 import type { SourceGraph } from "@/types/source-graph"
-
-export interface LocalProjectFile {
-  file: File
-  path: string
-}
-
-const PARSEABLE_EXTENSIONS = new Set([".js", ".jsx", ".rb", ".ts", ".tsx"])
-const GENERATED_PATH_SEGMENTS = new Set([".git", ".vite", "coverage", "dist", "node_modules"])
-const MAX_PROJECT_PATHS = 15000
-const MAX_PARSEABLE_FILES = 500
-const MAX_PARSEABLE_FILE_BYTES = 512 * 1024
-
-export interface ProjectImportPlan {
-  files: LocalProjectFile[]
-  skippedCount: number
-  parseableCount: number
-}
-
+import { planProjectImport, type LocalProjectFile } from "@/services/project-folder-service"
 
 const serverApplicationOrigin = import.meta.env.VITE_SERVER_APPLICATION_ORIGIN?.replace(/\/$/, "") ?? ""
 
@@ -47,31 +30,10 @@ export async function createGraphFromSourceFile(file: File): Promise<SourceGraph
   return responseBody.graph
 }
 
-export function planProjectImport(projectFiles: LocalProjectFile[]): ProjectImportPlan {
-  const files = projectFiles.filter((projectFile) => !isGeneratedPath(projectFile.path))
-  const parseableFiles = files.filter(isParseableProjectFile)
-
-  if (files.length > MAX_PROJECT_PATHS) {
-    throw new Error(`Dimension can map up to ${MAX_PROJECT_PATHS} project paths at once; ${files.length} were selected.`)
-  }
-
-  if (parseableFiles.length > MAX_PARSEABLE_FILES) {
-    throw new Error(
-      `Dimension can parse up to ${MAX_PARSEABLE_FILES} source files at once; ${parseableFiles.length} were selected after generated folders were skipped.`,
-    )
-  }
-
-  return {
-    files,
-    skippedCount: projectFiles.length - files.length,
-    parseableCount: parseableFiles.length,
-  }
-}
-
 export async function createGraphFromProjectFiles(rootName: string, projectFiles: LocalProjectFile[]): Promise<SourceGraph> {
   const plan = planProjectImport(projectFiles)
   const paths = plan.files.map((projectFile) => projectFile.path)
-  const parseableFiles = plan.files.filter(isParseableProjectFile)
+  const parseableFiles = plan.parseableFiles
   const formData = new FormData()
   formData.append("rootName", rootName)
   formData.append("paths", JSON.stringify(paths))
@@ -94,20 +56,6 @@ export async function createGraphFromProjectFiles(rootName: string, projectFiles
   }
 
   return responseBody.graph
-}
-
-function isParseableProjectFile(projectFile: LocalProjectFile): boolean {
-  return PARSEABLE_EXTENSIONS.has(extensionFor(projectFile.path)) && projectFile.file.size <= MAX_PARSEABLE_FILE_BYTES
-}
-
-function isGeneratedPath(path: string): boolean {
-  return path.split(/[\\/]+/).some((part) => GENERATED_PATH_SEGMENTS.has(part))
-}
-
-function extensionFor(path: string): string {
-  const match = /\.[^.\\/]+$/.exec(path)
-
-  return match?.[0] ?? ""
 }
 
 async function readGraphImportResponse(response: Response): Promise<GraphImportResponse> {
